+++ src/Service/DynamoDb/src/Exception/ConditionalCheckFailedException.php
@@ -3,6 +3,7 @@
 namespace AsyncAws\DynamoDb\Exception;
 
 use AsyncAws\Core\Exception\Http\ClientException;
+use AsyncAws\DynamoDb\ValueObject\AttributeValue;
 use Symfony\Contracts\HttpClient\ResponseInterface;
 
 /**
@@ -10,6 +11,21 @@ use Symfony\Contracts\HttpClient\ResponseInterface;
  */
 final class ConditionalCheckFailedException extends ClientException
 {
+    /**
+     * Item which caused the `ConditionalCheckFailedException`.
+     *
+     * @var array<string, AttributeValue>
+     */
+    private $item;
+
+    /**
+     * @return array<string, AttributeValue>
+     */
+    public function getItem(): array
+    {
+        return $this->item;
+    }
+
     protected function populateResult(ResponseInterface $response): void
     {
         $data = $response->toArray(false);
@@ -17,5 +33,19 @@ final class ConditionalCheckFailedException extends ClientException
         if (null !== $v = (isset($data['message']) ? (string) $data['message'] : null)) {
             $this->message = $v;
         }
+        $this->item = empty($data['Item']) ? [] : $this->populateResultAttributeMap($data['Item']);
+    }
+
+    /**
+     * @return array<string, AttributeValue>
+     */
+    private function populateResultAttributeMap(array $json): array
+    {
+        $items = [];
+        foreach ($json as $name => $value) {
+            $items[(string) $name] = AttributeValue::create($value);
+        }
+
+        return $items;
     }
 }