+++ src/Service/Lambda/src/LambdaClient.php
@@ -113,7 +113,7 @@ class LambdaClient extends AbstractApi
 
     /**
      * Deletes a Lambda function. To delete a specific function version, use the `Qualifier` parameter. Otherwise, all
-     * versions and aliases are deleted.
+     * versions and aliases are deleted. This doesn't require the user to have explicit permissions for DeleteAlias.
      *
      * To delete Lambda event source mappings that invoke a function, use DeleteEventSourceMapping. For Amazon Web Services
      * and resources that invoke your function directly, delete the trigger in the service where you originally configured