+++ src/Service/MediaConvert/src/ValueObject/CaptionSelector.php
@@ -14,8 +14,8 @@ final class CaptionSelector
     /**
      * The specific language to extract from source, using the ISO 639-2 or ISO 639-3 three-letter language code. If input
      * is SCTE-27, complete this field and/or PID to select the caption language to extract. If input is DVB-Sub and output
-     * is Burn-in or SMPTE-TT, complete this field and/or PID to select the caption language to extract. If input is DVB-Sub
-     * that is being passed through, omit this field (and PID field); there is no way to extract a specific language with
+     * is Burn-in, complete this field and/or PID to select the caption language to extract. If input is DVB-Sub that is
+     * being passed through, omit this field (and PID field); there is no way to extract a specific language with
      * pass-through captions.
      *
      * @var string|null
@@ -24,9 +24,9 @@ final class CaptionSelector
 
     /**
      * The specific language to extract from source. If input is SCTE-27, complete this field and/or PID to select the
-     * caption language to extract. If input is DVB-Sub and output is Burn-in or SMPTE-TT, complete this field and/or PID to
-     * select the caption language to extract. If input is DVB-Sub that is being passed through, omit this field (and PID
-     * field); there is no way to extract a specific language with pass-through captions.
+     * caption language to extract. If input is DVB-Sub and output is Burn-in, complete this field and/or PID to select the
+     * caption language to extract. If input is DVB-Sub that is being passed through, omit this field (and PID field); there
+     * is no way to extract a specific language with pass-through captions.
      *
      * @var LanguageCode::*|null
      */