+++ src/Service/MediaConvert/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - AWS api-change: This release introduces the bandwidth reduction filter for the HEVC encoder, increases the limits of outputs per job, and updates support for the Nagra SDK to version 1.14.7.
+- AWS enhancement: Documentation updates.
 
 ## 0.1.0